requirements.txt
asttokens==2.4.1
certifi==2024.7.4; python_version >= '3.6'
colorama==0.4.6; sys_platform == 'win32'
comm==0.2.2; python_version >= '3.8'
contourpy==1.2.1; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
debugpy==1.8.5; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
executing==2.0.1; python_version >= '3.5'
fonttools==4.53.1; python_version >= '3.8'
geopandas==1.0.1; python_version >= '3.9'
imageio==2.35.0; python_version >= '3.8'
ipykernel==6.29.5; python_version >= '3.8'
ipython==8.26.0; python_version >= '3.10'
jedi==0.19.1; python_version >= '3.6'
joblib==1.4.2; python_version >= '3.8'
jupyter-client==8.6.2; python_version >= '3.8'
jupyter-core==5.7.2; python_version >= '3.8'
kiwisolver==1.4.5; python_version >= '3.7'
lazy-loader==0.4; python_version >= '3.7'
matplotlib==3.9.2; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.3; python_version >= '3.10'
numpy==1.26.4; python_version >= '3.9'
opencv-python==4.10.0.84; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
parso==0.8.4; python_version >= '3.6'
pillow==10.4.0; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psutil==6.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pure-eval==0.2.3
pygments==2.18.0; python_version >= '3.8'
pyogrio==0.9.0; python_version >= '3.8'
pyparsing==3.1.2; python_full_version >= '3.6.8'
pyproj==3.6.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pywin32==306; sys_platform == 'win32' and platform_python_implementation != 'PyPy'
pyzmq==26.1.0; python_version >= '3.7'
scikit-image==0.24.0; python_version >= '3.9'
scipy==1.14.0; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
shapely==2.0.5; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.6.3
tifffile==2024.8.10; python_version >= '3.9'
tornado==6.4.1; python_version >= '3.8'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.12'
tzdata==2024.1; python_version >= '2'
wcwidth==0.2.13